Restructure WNetwork
Restructure WNetwork to avoid the use of select! by placing both the sink and stream of a websocket
connection inside of a single task, and using a handle to communicate with the task.

Use a stream combinator to combine the input and output streams to allow a single asynchronous loop
to take care of both streams.

This will allow the keepalive timers to be managed in a single task, for all handles, avoiding
issues around having to fuse a timeout future.

This obsoletes #12, as the structure no longer relies on having a timer future.

Closes #12