Change StakeTable from HashMap to BTreeMap
The implementation of DynamicCommittee will select a different
leader if it iterates over the stake table in a different order.
The iteration order for HashMap is unspecified and arbitrary, and
I have observed different nodes choosing different leaders with
this implementation.

With few enough nodes, this can lead to a situation where not enough
nodes agree on any one leader to form a QC, resulting in loss of
progress. This appears to be the root cause of the cli_integration_tests
failure in Espresso.

BTreeMap has a stable, deterministic iteration order, so it solves
this issue.