feat!: Remove tc and switch to list-of-signatures
This commit removes the use of the threshold_crypto crate, and changes
the quorum certificate format to be a list of full signatures instead of
a single threshold signature. A demonstration implemenation of the
SignatureKey trait using the ed25519-compact crate as a backend is
provided.

The following supporting changes are also made:
 - An Implementation of `SignatureKey` for types providing an ed25519
   implementation
 - Adds a SigningKey type to the NodeImplementation trait
 - Adds a signing_key field to PhaseLockInner, and propagates the change
   through the constructor
 - Optimize ed25519-compact in the dev profile
 - Have PhaseLock store the public key set, allowing it to check QCs for validity
 - Make PubKey generic in key type
 - Removal of the old committee.rs file, as it relied heavily on
   threshold_crypto
 - Removal of explicit dependencies on arkworks crates
 - Update the dentry-simulator example to use the new key/quorum format
 - Update the multi-machine example to use the new key/quorum format
 - Update the tests to use the new key/quorum format

This leaves open a number of TODOs and FIXMEs that are tracked in #170

BREAKING CHANGE: Adds new type paramaters and subsantially modifies APIs
for phaslock types