feat!: Remove Threshold Cryptography
Remove the use of threshold cryptography from phaselock.

This results in the removal of the PubKey and PrivKey types, which are
now replaced with generics based on the SignatureKey trait, for which a
demo implementation based on ed25519 is provided, which is now
integreated into the NodeImpelmentation trait.

The ed25519 implementation, furthermore, implements the new
TestableSignatureKey trait, which provides an interface for
deterministic key generation inside of unit and integration tests.

The signature key trait is made to work with raw buffers of bytes, to
avoid baking in assumptions about signature formats, provided through
type safety wrappers.

The quorum certificate format has been changed to use a list of
signatures and the public keys that generated them, rather than a single
threshold signature. As this requires knowledge of the list of allowed
keys to properly verify, the verification of quorum certificates has
been moved into the ConsensusApi trait.

This change completely removes threshold_crypto from phaselock's
dependencies, as well as removes the old DynamicComittee implementation,
as it is totally incompatible with these changes, also clearing the
arkworks deps from the list.

BREAKING CHANGE: Removes PubKey and PrivKey types, adds new generic to
many traits