Fix computation of leaf commitment in replica.
Previously, the replica assumed that the parent of the new
proposal was its own `high_qc`. This is incorrect, since nothing
guarantees that the replica will have the same `high_qc` as the
leader which created the proposal.

Instead, we make sequencing consensus act a bit more like
validating consensus -- we take the leaf commitment from the
proposals `justify_qc` as the parent, get the corresponding leaf
from storage, and validate `justify_qc` to ensure the leader is
being honest.

There is an exception for the genesis view, since the parent of
the genesis QC is not a real leaf. In this case, we look up the
genesis leaf explicitly from storage, which matches the logic the
leader uses to construct the proposal. This edge case does not
affect validating consensus because it is handled by the leader,
and rather than handle it themselves, the replicas get the parent
commitment directly from the proposal. But in sequencing consensus,
there is no explicit parent commitment in the proposal.

This change poses a problem for the DA committee members, who are
no longer able to construct a leaf, since the DA proposal does not
include a justify QC _or_ parent commitment. This is OK, since the
DA members only needed to construct a leaf so they could store at
as a hacky way of associating a leaf with a full block (since only
the DA member has the full block). We will replace this in a future
commit with a separate store for blocks only -- not leaves -- so
the DA member can store its block without constructing a leaf.