Clean up header genesis
Previously, `Header::genesis` was required to also return a payload
and metadata. This is redundant, because we also have `Payload::genesis`.
In addition, `Header::genesis` was expected to compute a genesis VID
commitment, since a VID commitment is required to construct a header.
This breaks abstraction, since in all other cases only HotShot computes
VID stuff, never the application. It is also error prone: I am adding a
feature to the query service to handle the genesis VID case, where we
don't receive VID data from HotShot. This feature requires that the
genesis VID is computed uniformly regardless of the application. Currently,
we use 8 storage nodes for this commitment in the HotShot example types
(quite arbitrarily) and 1 in the sequencer types. This change standardizes
the computation of the genesis VID commitment.