Make metrics labels more compatible with Prometheus
* Add `MetricsFamily` trait to group related metrics that are
  partitioned by values of their labels. This gives us a way to create
  metrics that actually use Prometheus labels, a capability we haven't
  had until now.
* Rename `Label` to `Text` to distinguish it from the Prometheus
  concept of a label
* Change the way `Text` works. Prometheus doesn't support top-level
  key-value pairs with text values, so implementing the old `Label`
  trait for Prometheus, we were forced to hack it by putting the value
  in a comment. A better approach is to have a "text" metric where the
  value of the metric is the name (this can be implemented, e.g., as a
  gauge with a dummy value of 1). Additional key-value pairs can be
  added using families/labels as described above.
* Remove dead code `Metrics` implementations in consensus and networking

The motivation for this is a request from node operators to make the
version metrics populated by the sequencer more parseable. For this,
we only need a family with a single instance, whose label values contain
information like git revision and whatnot. However, this implementation
is more general, and will allow us to fully take advantage of Prometheus
labels if we want to in the future.

Examples of both the simple "version label" usage and a more complex use
of labels can be found in the doc comments.