CI: compile without lockfile (#3850)
Closes #3849

Since hotshot is a library I think this check is quite important because
(I think) if it fails this means that anyone starting a new project with, or
adding hotshot as new dependency to a project will have the build break
with likely annoying/difficult to resolve dependency issues.

I was going to use just first and noticed the installation in the CI can be
simplified. However since the just file does set env vars that are used by
the rust-cache action (after the action is initialized and the cache key is
computed) I opted for not using just for the new CI job.