Add back in consideration for `blocks_per_epoch` of 0
With the previous change the explicit check for `blocks_per_epoch`
being zero was removed, thereby removing the protection of division
by 0.  This check should be re-added making this change somewhat less valuable as we're unable to remove all conditionals.

This change has been made to re-add the initial conditional in order to ensure compatibility with the previous implementation.